3bduu

+++ include/cub3D.h
@@ -7,8 +7,8 @@
 # include <fcntl.h>
 # include <stdio.h>
 #include "../libft/libft.h"
-#include<mlx.h>
-// #include "../minilibx-linux/mlx.h"
+// #include<mlx.h>
+#include "../minilibx-linux/mlx.h"
 #include <math.h>
 //# include "./get_next/get_next_line.h"
 
@@ -147,7 +147,7 @@ void minimap(t_my_map *src);
 int     player_press(int keycode,t_my_map *src);
 int     player_up(int keycode,t_my_map *src);
 void	my_mlx_pixel_put(t_data *data, int x, int y, int color);
-void    draw_square(t_my_map *src,int start_h,int start_w,int end_h,int end_w,int color);
+void draw_square(t_my_map *src,float start_h,float start_w,float end_h,float end_w,int color);
 void    draw_circle(t_my_map *src,int posX,int posY,int ray);
 int mapHasWallAt(t_my_map *src,float x, float y) ;
 #endif
